# swagger.yml
swagger: '2.0'
schemes:
  - http
host: 'localhost:5000'
info:
  description: RESTful API for Ostadkar Backend
  version: 0.0.1
  title: Ostadkar REST API
  contact:
    email: na.farahzadi@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /zones:
    get:
      tags:
        - operators
        - providers
      summary: List Zones
      operationId: listZones
      description: Returns list of zones available
      produces:
        - application/json
        - application/msgpack
        - application/cbor
      parameters:
        - in: header
          name: Accept
          type: string
          required: false
      responses:
        '200':
          description: list of zones with attributes
          schema:
            type: array
            items:
              $ref: '#/definitions/ZoneInfo'
        '404':
          description: No zone found
    post:
      tags:
        - operators
      summary: Registers new zone
      operationId: registerZone
      description: Registers an zone with its data
      consumes:
        - application/json
        - application/msgpack
        - application/cbor
      produces:
        - application/json
        - application/msgpack
        - application/cbor
      parameters:
        - in: body
          required: true
          name: zoneInfo
          schema:
            $ref: '#/definitions/ZoneInfo'
      responses:
        '201':
          description: Zone registered successfully
          schema:
            type: object
            properties:
              sid:
                type: string
        '400':
          description: Zone data invalid
          schema:
            $ref: '#/definitions/ErrorInfo'
  '/zones/{slug}':
    get:
      tags:
        - customers
        - providers
        - operators
      summary: Gets some specific zone details
      operationId: getZoneInfo
      description: Gets a tracking code and return corresponding zone details
      produces:
        - application/json
        - application/msgpack
        - application/cbor
      parameters:
        - in: path
          required: true
          name: slug
          type: string
      responses:
        '200':
          description: Returns zone with specified slug
          schema:
            $ref: '#/definitions/ZoneInfo'
        '404':
          description: No zone exists with this slug
          schema:
            $ref: '#/definitions/ErrorInfo'
    put:
      tags:
        - operators
      summary: Modifies speicified skill information
      operationId: editZoneInfo
      description: 'Modifies sefvice name, description and attributes.'
      produces:
        - application/json
        - application/msgpack
        - application/cbor
      consumes:
        - application/json
        - application/msgpack
        - application/cbor
      parameters:
        - in: path
          name: slug
          required: true
          type: string
        - in: body
          name: skillInfo
          required: true
          schema:
            $ref: '#/definitions/ZoneInfo'
      responses:
        '204':
          description: Successfully modified the skill
        '400':
          description: Invalid skill data
          schema:
            $ref: '#/definitions/ErrorInfo'
        '404':
          description: There is no skill with this slug
    delete:
      tags:
        - operators
      summary: Disables a single skill specified by slug
      operationId: disableZone
      description: Cancels a specific skill with slug
      parameters:
        - in: path
          name: slug
          required: true
          type: string
      responses:
        '204':
          description: Successfully disabled the skill
        '404':
          description: The is no skill with this slug.
          schema:
            $ref: '#/definitions/ErrorInfo'
        '406':
          description: This skill could not be disabled.
          schema:
            $ref: '#/definitions/ErrorInfo'
...
definitions:
  ErrorInfo:
    title: error
    description: Error description
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: number
        format: integer
      message:
        type: string
      uri:
        type: string
...
  ZoneInfo:
    description: Provider zone information
    title: zone
    type: object
    required:
      - id
      - city_slug
      - slug
      - title
      - created_at
      - is_active
    properties:
      id:
        type: number
        description: Unique identifier of the zone
        format: int32
        example: 123
      city_slug:
        type: string
        maxLength: 25
        description: Alphanumeric url-valid string for natural accessing the item
        example: tehran
      city_title:
        type: string
        maxLength: 100
        description: Title of the city
        example: تهران
      slug:
        type: string
        maxLength: 25
        description: Alphanumeric url-valid string for natural accessing the item
        example: pasdaran
      title:
        type: string
        maxLength: 100
        description: Title of the zone
        example: پاسداران
      created_at:
        type: string
        format: date-time
        example: '2018-09-05T23:11'
      updated_at:
        type: string
        format: date-time
        example: '2018-09-06T23:11'
      is_active:
        type: boolean
        example: true
...